:art: #35 moved platform related code into dedicated package
